% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_files.R
\name{fars_read}
\alias{fars_read}
\title{Read fars csv data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character; name of the file.}
}
\value{
a tibble containing the fars data.
}
\description{
Read the fars data file in csv format. Return an error if the file does not
exist.
}
\examples{
\dontrun{
fars_read(fars_data.csv)
}
}
