% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_files.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read the fars data from different years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector or list; the years that will be read.}
}
\value{
a list of tibbles containing the data for the years.
}
\description{
For each year passed as argument, read the fars data, create a column with
the year and select the month and the year. Return an error if the year is
not valid. The files must be in the working directory.
}
\examples{
fars_read_years(c(2014, 2015))

}
