% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_analysis.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Sumarize years of fars data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector or list; the years that will be read.}
}
\value{
a tibble containing the number of fars by month and year.
}
\description{
Create a table containing the the number of fars where the months are
represented as rows and the years as columns.  The files must be in the
working directory.
}
\examples{
fars_summarize_years(c(2014, 2015))

}
