% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_files.R
\name{make_filename}
\alias{make_filename}
\title{Create the file name.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year that will appear in the name of the file.}
}
\value{
a character representing the name of the file.
}
\description{
Create the file name from the year passed to the function using the format
accident_year.csv.bz2.
}
\examples{
\dontrun{
make_filename(2014)
}
}
